requirements.txt
Flask==3.1.2
flask_cors==6.0.2
flask_talisman==1.1.0
python-dotenv==1.2.1
Requests==2.32.5
